Moved some helper methods to ancillary files
